Removed an MPI_Barrier.
Renamed 'sol' to 'x' which is later reused to output the solution.

Define the vector 'X' and its initial guess for the pcg by
ParallelAverage; this initialization sets the correct values at the
essential dofs -- this is the effect of the ParallelAssembles of 'a'
and 'b' at the essential dofs.