Added new FiniteElementCollection sub-classes: H1_FECollection, L2_FECollection, and RT_FECollection for arbitrary high-order FE spaces. These are based on a number of new FiniteElement sub-classes H1_*, L2_*, and RT_* elements of arbitrary order on all types of reference elements.
Fixed a bug in DofOrderForOrientation for quads in CubicFECollection
and RT1_3DFECollection.

Reimplemented GridFunction::ProjectBdrCoefficientNormal to work with
the new RT_FECollection.

In the GridFunction::Compute*Error methods, increased the default
quadrature rule order to (2*FE_Order + 1).

In class GridFunction, added variations of the methods
ProjectBdrCoefficient and ComputeL2Error that take 'Coefficient &' as
input (the old versions take 'Coefficient *[]').

Added new Geometry::Name variable defining the names of the base
geometries.

In class GeometryRefiner, added a new parameter 'type' which can be
set to generate non-uniform subdivisions based on the non-uniform 1D
subdivision defined by Poly_1D::ClosedPoints (the Gauss-Lobatto points
by default).

Added new method DenseMatrix::TestInversion that inverts a DenseMatrix
and prints information on the numerical conditioning of the inversion.