Added support for NURBS meshes including NURBS-specific refinement functions: KnotInsert and DegreeElevate.
Added an optional argument to Mesh::GetFaceElementTransformations
method to specify which data fields should be returned.

In ParMesh::ParMesh, re-enumerate the local vertices to preserve the
global ordering, assign interior boundary elements on processor
interface to only one of the two neighbors.

The partitioning parameter of ParMesh::ParMesh will be deleted inside
the constructor or stored in the ParNURBSExtension for later deletion.