Extended the reading of VTK input to directly support meshes that have been exported from VisIt.
With this change, we can convert and use any mesh that can be opened in VisIt
(e.g. from a silo file). Note however, that VisIt uses floats (as opposed to
doubles) for the mesh coordinates, so there may be some loss of accuracy. One
also needs to be careful with the material attributes, which may need to be
shifted if they start from 0.

The export procedure is as follows:

1) Open the mesh in VisIt.

2) Using Control/Expressions define a new scalar mesh variable with definition
   matvf(material,1).

3) Draw a pseudocolor plot of the new variable.

4) Select File / Export Database, Variables: Meshes, Filetype: VTK.