Added new abstract base class NonlinearForm representing general nonlinear finite element operators defined locally by classes derived from the new base class NonlinearFormIntegrator.
NonlinearForm defines assemble-based action (Mult), as well as assembly of the
gradient operator when defined in its integrators enabling inversion with
Newton iteration.

Parallel version is also implemented in the class ParNonlinearForm.

The class NonlinearFormIntegrator is now the base class for class
BilinearFormIntegrator inheriting the matrix-free action. Currently, only
DiffusionIntegrator implements such an action.

Added DGTraceIntegrator for integrating the form
    alpha < (u.n) {v},[w] > + beta < |u.n| [v],[w] >
and the related LinearFormIntegrator sub-class BoundaryFlowIntegrator
    (alpha/2) < (u.n) f, w > - beta < |u.n| f, w >.

Adding a new integrator class, GroupConvectionIntegrator, implementing a
version of the so called group finite element formulation of Fletcher.

Some tweaks in the implementation of class ConvectionIntegrator.