Improved the implementations of DenseMatrix::CalcEigenvalues and DenseMatrix::CalcSingularvalue (2x2 and 3x3 matrices).
In class DenseTensor, added methods for direct entry access using A(i,j,k).
Also added a method that performs the action of the assembled matrix if the
DenseTensor contains the local element matrices.

Added a global function CheckFinite that counts the NaNs and +/-Infs
in a double array. Also, in class Vector and DenseMatrix added methods
with the same name that check their entries.

Added a method DenseMatrix::MultTranspose with pointer arguments and a method
DenseMatrix::MaxMaxNorm that computes the norm ||A|| = max_{ij} |A_{ij}|.
Also, optimized the implementations of some DenseMatrix functions.