Adding an option to build the experimental code in dev/ for profiling on bgq. To use this option build mfem using:
  make CC=bgxlC_r OPTIM_OPTS="-O2 -qmaxmem=-1 -qnoeh"

and then build the code in dev/ using:

  make ENABLE_BGQPROF=YES

With this build the code can be run with 64 mpi tasks on one node to evaluate
performance when the node is fully loaded. The is just the same serial code run
on multiple mpi tasks. The problem size is adjusted to fit all the data in the
16G on the node.