Updated header file to explicitly include other headers they need.
In the FiniteElement class, added a new member, MapType that defines the way
basis functions are mapped from the reference element to any element. The
supported MapTypes are given by an anonymous enum in FiniteElement:
{ VALUE, INTEGRAL, H_DIV, H_CURL } -- see the doxygen comment for description.

In the NodalFiniteElement class, added a method SetMapType -- this can be used
to switch the MapType (VALUE <--> INTEGRAL) for any NodalFiniteElement; modified
the methods defined in NodalFiniteElement to support the new MapType - INTEGRAL.

For VectorFiniteElements, MapType is set accordingly and can not be changed.

Changed the class from which the H1Pos_* finite elements inherit from -- now
they inherit from PositiveFiniteElement -- this way they inherit the general
implementations of the Project methods.

Deleted the L2Vol_SegmentElement, this can be replaced by standard
L2_SegmentElement where the MapType is set to INTEGRAL.

Fixed a bug in ND_TriangleElement::CalcVShape.

In the RT_FECollection and ND_FECollection, use the correct MapType for the
L2 trace spaces.

In class FiniteElementSpace:
 * Made the method NC_GlobalRestrictionMatrix protected.
 * Changed the method GlobalRestrictionMatrix to work with nonconforming meshes
   by calling NC_GlobalRestrictionMatrix.
 * Small bugfix to support RT_* and ND_* spaces in NC_GlobalRestrictionMatrix.
 * In a nonconforming space, if there are no slave dofs, cP and cR are set to be
   NULL since the space is actually conforming.

In class Mesh, use the NCMesh::coarse_elements only when WantTwoLevelState is
set.

In class NCMesh:
 * In the methods ProcessMasterEdge, ConstrainEdge, ProcessMasterFace, and
   ConstrainFace check for empty dof-arrays -- this is a bugfix and a small
   optimization as well.
 * In the method GetInterpolation, return NULL matrices when there are no slave
   dofs.