Changes related the DPG:  * In class NT_FECollection replace L2Vol_SegmentElement with "INTEGRAL"    L2_SegmentElement.  * Minor changes in ex8* examples.
Changes not related to DPG:

In class Operator:
 * Modified the class to have two separate sizes: 'height' and 'width' for the
   output and input sizes, respectively. These can be accessed using the Height
   (or the synonym NumRows) and Width (or the synonym NumCols) methods.
 * Removed the 'size' member and the Size method.
 * Added a general constructor for different height and width.

Updated the implementations of classes derived from Operator to account for the
changes in the base class.

Defined method a Size in the classes DenseMatrix (returns width as before),
SparseMatrix (returns height as before), DenseMatrixInverse (square matrix),
BilinearForm (square matrix) -- for backward compatibility and where it makes
sense (for square operators).

Removed the use of the Size method of Operator and derived classes from the
library and the examples.

Updated GridFunction::GetVectorGradientHat and methods that use it for the case
where there are different mesh/space/vector dimensions.

Modified some parameter names in constructors of HypreParMatrix.

Added a shortcut for Vector::StealData that returns the data pointer.