Rewrite of the makefile "build system":
Added new target 'config' for configuring MFEM:
   make config [OPTIONS] ...
The supported OPTIONS are given in the file 'config/user.mk.in'; this file can
be copied to 'config/user.mk' and edited to customize the config options instead
of specifying them on the command line. The file 'config/user.mk' can be
overwritten with the USER_CONFIG=/path/to/file option.

Running 'make config' overwrites the files 'config.hpp' and 'config/config.mk'
with the specified options. The latter file can be used by other makefiles
(e.g. it is used by 'examples/makefile') to facilitate compiling and linking
with the MFEM library.

Added new target 'info' that displays the configured options.

After configuration, MFEM can be built by running 'make' or 'make -j <np>'.

Shortcut targets that configure and then build MFEM are similar to the old ones:
'serial', 'parallel', 'debug', and 'pdebug'.

To build the examples use 'make' or 'make -j <np>' in the examples directory.
Based on the current MFEM configuration this will build the serial or the
parallel examples using the same config options as the library.