Fixed an issue in Mesh::HexUniformRefinement, where (in debug mode) the method CheckBdrElementOrientation was called to fix boundary element orientations after the 'bel_to_edge' Table was built.
Modified all calls to CheckBdrElementOrientation in debug mode to use 'false'
parameter (i.e. do not fix orientations) -- this way debug and optimized builds
should be identical.

Made the arrays Hexahedron::edges and Hexahedron::faces public. Simplified
some implementations using Hexahedron::edges and Mesh::hex_faces (which is the
same as Hexahedron::faces).

Added a protected method Mesh::FaceIsTrueInterior that in parallel checks for
shared faces in addition to (locally) interior faces. This is achieved by
checking faces_info[face_id].Elem2Inf -- it is now initialized to -1 and for
shared face, its value is overwritten in ParMesh::ExchangeFaceNbrData.

Fixed an issue in Mesh::GetBdrFaceTransformations, where in parallel, we did not
return NULL for shared faces -- now this is done using Mesh::FaceIsTrueInterior.

In the ParMesh constructor, assign boundary faces that are shared to the
processor for which the boundary element is oriented correctly (if possible).

In ParMesh::Print, added a local variable 'print_shared' to control whether the
shared boundaries are printed. Should we make this a parameter of the method?

Added some debug checks in mesh.cpp.