Add a the method ParBilinearForm::ParAddMult(const Vector & x,Vector & y, const int a); Expose the private method ParBilinearForm::ParallelAssemble(SparseMatrix * m);
The method ParBilinearForm::ParAddMult takes as paramenter two vectors, x and y, of size TrueDof and
set y += A*x, where A is the "true" operator contained in the Parallel Bilinear Form.
To do that:
- we distribute x --> x_proc;
- we perform on each processor the multiplication tmp_proc = A_proc * x_proc;
- we globally assembe tmp_proc --> tmp;
- we add y += a*tmp.
If one follows the convenction in NonlinearForm we should have called this method
simplu AddMult. However in order not to overload the method AddMult already
present in BilinearForm, I decided to call it ParAddMult.

The method ParBilinearForm::ParallelAssemble(SparseMatrix *m) was originally private
however it could be handy to have this method public so that it is possible
to call ParallelAssemble on a matrix m computed outside the BilinearForm object.
(see ex10p).