Running ex10p with implicit integrators exposes a memory error (Mismatched free() / delete / delete []) in HypreSmoother.
The problem appears in line hypre.cpp:1068 (HypreSmoother::SetOperator) and in line hypre.cpp:1249 (HypreSmoother::~HypreSmoother).
The problematic instruction (in both points of the code) is
   if (l1_norms)
      delete [] l1_norms;

The problem was that the way the memory is allocated depends on the type of smoother.

-- if smoother type is 1,2,3,4 we use HYPRE to allocate the memory for
   l1_norms (see hypre.cpp:1076 -> hypre_ParCSRComputeL1Norms(*A, type, NULL, &l1_norms); ),
   so we should use HYPRE_TFree(l1_norms);

-- if smoother type is 5 we use MFEM (new []) to allocate the memory for
   l1_norms (see hypre.cpp:1083 -> l1_norms = diag.StealData(); )
   se we should use delete[] l1_norms;

My solution was to always use Hypre to allocate the memory for l1_norms:
+      l1_norms = hypre_CTAlloc(double, height);
+      Vector ones(height), diag(l1_norms, height);

so that it is safe to always use hypre_TFree(l1_norms).