Support a wider variety of material descriptions in VTK mesh files.
With this change we can export any quadrilateral mesh from VisIt to MFEM. The
procedure is:

1) Using Control/Expressions define a new scalar mesh variable named "material"
   with definition: floor(dominant_mat(Materials)).

2) Draw the mesh and a pseudocolor plot of the "material" variable

3) Select File / Export Database,
   Variables: mesh material, Filetype: ITAPS_MOAB, Format: VTK

The resulting file should be in mfem-readable format with attributes marked
according to the VisIt materials.

Here are the details from the VTK file format document:

"""
Scalar definition includes specification of a lookup table. The definition of a
lookup table is optional. If not specified, the default VTK table will be used
(and tableName should be “default”). Also note that the numComp variable is
optional—by default the number of components is equal to one. (The parameter
numComp must range between (1,4) inclusive; in versions of VTK prior to vtk2.3
this parameter was not supported.)

SCALARS dataName dataType numComp
LOOKUP_TABLE tableName
s0
s1
...
sn-1
"""

See http://www.vtk.org/VTK/img/file-formats.pdf, page 5.