Added code to allocate and assemble matrices to support static condensation.  The class will either build 'mat' and 'mat_e' or the new matrices 'mat_ee', 'mat_ep', 'mat_rr', 'mat_pp', 'mat_pp_inv', and optionally 'mat_pe'.  The choice is based upon whether or not the FiniteElementSpace object was constructed to support static condensation.
Currently, the matrices are only allocated and assembled.  They cannot yet be used.  In fact much of the class still assumes 'mat' and 'mat_e' are created so many methods will result in segmentation faults.  However, the current code has passed initial tests confirming that the static condensation matrices are being properly built.  They agree with the matrices produced by the EPBilinearForm class with and without precomputing the sparsity pattern.

Another caveat is that the current implementation does not support vdim>1.  This has to do with the assumption that private DoFs will be at the end of the local vectors.  When vdim>1 this will not be true if the DoF ordering is set to Ordering::byNODES (which is the default).  Resolving this will likely involve changes to FiniteElementSpace and GridFunction (and probably other classes).