The BilinearForm class should now be ready to replace the EPBilinearForm.  Some of the methods have slightly different names but the functionality should be the same.
Note that elimination of essential boundary conditions has not been altered and so it won't work as is with static condensation.  This is only temporary.  I plan to incorporate the code that Jakub has been working on to eliminate essential BCs but I'm not ready to look at that yet.  It sounds like it only requires small changes but I want to get this version of static condensation working in parallel first before adding support for BCs.