Updates to class DenseMatrix: * Added data member 'capacity' that stores the size of the dynamically allocated   data array. If 'capacity' is zero or negative, then DenseMatrix does not own   the data array and abs(capacity) is the actual capacity. * When resizing the DenseMatrix with SetSize, a new data array is dynamically   allocated only if the current capacity is not sufficient to hold the new size.
Updated the implementation of some operations on DenseMatrixes; in particular
the matrix-matrix product functions Mult and AddMult are better suited for
simple auto-vectorization -- I get some nice speedups with gcc 5.

Added new class LUFactors that can compute LU factorization (with pivoting) and
perform various operations with the factored data, e.g. apply the factors to a
vector/matrix, apply the inverse matrix to a vector/matrix, compute 2x2 block
factorizations (Schur complements), compute the inverse matrix from the factors,
etc.

Updates to class DenseMatrixInverse:
* In the implementation, use the new class LUFactors.
* Added new methods: Mult (multiply the inverse matrix by another matrix) and
  GetInverseMatrix.

In the files fe.hpp and fe.cpp: instead of using the method DenseMatrix::Invert
to compute and store the inverse of a matrix, use a DenseMatrixInverse to
compute LU factorization and use that to perform the action of the inverse
matrix. This improves the accuracy degradation due to round-off for high-order
finite elements on triangles and tets (they use matrix inversion for change of
basis).