This is a first draft of a class to wrap the LOBPCG eigenvalue solver as it exists in hypre.  The new example, ex11p, is a modified version of ex1p which computes and displays the eigenmodes of the Laplacian operator in 2D or 3D.
I will still need to extend or modify these new classes because I will need to find the eigenmodes of systems defined as BlockOperators rather than simply HypreParMatrices.  I have a test code which solves such a system so I know it can be done but I have yet to figure out an elegant way to support this capability.