Added support for a discretization-enhanced version of BoomerAMG designed specifically for linear elasticity problems, as described in:
Baker, Kolev, Yang, "Improving algebraic multigrid interpolation operators for
linear elasticity problems", NLAA 2009, DOI:10.1002/nla.688

This is implemented as an optional SetElasticityOptions() method of the
HypreBoomerAMG class, similar/alternative to SetSystemsOptions() which
corresponds to the "unknown" approach in the paper.

Example 2p, now has a new parameter "-elast" that switches to this solver (the
old default solver is "-sys"). There is definite (though not great) improvement
in the PCG iterations, including in parallel and at higher orders (I tested up
to 4th.)

This solver has a lot of additional options, for which I have picked some
default values based on my testing. We could adjust those in the future.