Fix internal macro naming conflict.
_MFEM_MESSAGE created an internal variable name "s".  On line 63 of
densemat.cpp, we tried writing out a variable called "s" to this stream,
which was hidden.  Essentially the macro expaned to

  s << s;

In C++98, the second s was implicitly converted to a bool, and written out.
This implicit conversion has been removed in C++11/14, exactly because
bugs like this existed.  Here's the proposal the removed that operation

http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2007/n2333.html

So I've change the internal name from "s" to "mfemMsgStream".