GitHub Pull Request #29 by @martemyev:
Fixed a possible bug in VectorMassIntegrator::AssembleElementMatrix

VectorMassIntegrator::AssembleElementMatrix got an information about the
dimension of the vector problem based on the dimension of the vector or matrix
coefficient. If the integrator is supposed to use a scalar coefficient the
dimension was equal to the dimension of the finite element, which is okay if
it's a domain integration, but is wrong when it's integration on a boundary.