Added documentation for the eigensolver examples, including eigensolver choices in the examples README file.
It was not easy to pick screenshots for these examples to put in the README
file. Many of them look really cool, and I just selected the first ones that
looked interesting to me. If you think you have a cooler picture, please feel
free to replace it. (Even if you don't want to put it in the README, please save
it for the Gallery on the web site.)