A few small fix-ups to make examples 3/3p and 4/4p work on surface meshes (as well as a few minor updates in the CHANGELOG).
The curl-curl and grad-div linear systems on surface meshes seem to be correct,
in particular because the AMS and ADS solvers seem to converge fine on them in
parallel. We don't compute the error correctly, and we can't visualize the
vector field on the surface, but when running on a flat version of
start-surf.mesh we get the same error as when running on star.mesh.