The main thing to note here is the minimal changes needed to mesh.cpp to support non-orientable surface meshes.
The two mesh files being added are flat  meshes (the Klien bottle mesh is simply a 2D mesh) which have the topology of a Mobius strip and a Klien bottle.  The "transform-mesh.cpp" program can be used to project the flat mobius-strip mesh into 3D.  It might be useful to expand upon this program to support other parametrized surfaces.  It would be ideal if we could figure out a way to encode the parameterization into the mesh files themselves but this might be too much to ask for.