Improvements to the GLVis visualization in the tesla and volta miniapps.
This still has some robustness issues in parallel, but is a significant
improvement over the previous state on my Linux box. I think it will be really
cool if we can get this working well so our screens can look like those of stock
traders :).

The newest version of GLVis (supporting the window_geometry socket command) is
required.