This is a modified version of ex6p intended to mimic the way a time evolution code might use refinement and derefinement.
I'm committing to code to start a discussion about how such a code should be written.  I expect this is a rather naive implementation (even ignoring the fact that it refines and derefines at every time step which a realistic code would certainly not do.)

Running the code with its default options leads to a refinement which appears far from optimal and even counter intuitive.  Could this be a limitation of the error estimator or is it due to the naive implementation of the refinement/derefinement strategy?