Cleaned up the refinement/derefinement interface in Mesh a bit:
There is no Mesh::GeneralDerefinement (that's nontrivial in parallel
due to synchronization issues), the method is now called Mesh::DerefineByError
which is more accurate. There is one more overload of DerefineByError
which takes an error vector.

Analogously, you can now also RefineByError, which is useful in ex6 and
similar and is a counterpart to DerefineByError. Of course RefineByError
is very simple and uses GeneralRefinement internally. Again there is an
overload for an error vector.