Cleaned up the predefined solutions in ex15p and updated some comments.
The exact solution is now prescribed on the Dirichlet boundary, so
there are no weird effects then the fronts reach the boundary.

There are two problem setups: "front" and "ball". Both work in 2D and 3D now.

I modified the "ball" setup so that an actual ball step function
is calculated which gives sharper gradients that can be tracked.