In class Mesh, add method GetBdrElementAdjacentElement.
Made NCMesh::Node::UnrefEdge a static method with a first parameter Node*;
this fixes a warning about the check "this != NULL".

In class Ordering, add two static methods: Map and DofsToVDofs with a template
parameter Ordering::Type and use them to implement relevant methods in class
FiniteElementSpace and a few other places.

Add a bool parameter to ParMesh::GetSharedFaceTransformations (with a default
value = true) that controls if the data for side "2" is returned.

In class Hybridization, add methods AssembleBdrMatrix and Reset.

Add 'sequence' member to BilinearForm, used in the Update method.

Add some comments.