Move final elmat computation to AssembleFaceMatrix
In both AssembleBoundaryFaceMatrix and AssembleInteriorFaceMatrix the final
computation of elmat matrix was the same: elmat := -elmat + sigma*elmat^t + jmat
Therefore, to avoid repeating the same code, this part was moved to the caller
method AssembleFaceMatrix. To make this work, the jmat matrix had to become a
member of the DGElasticityIntegrator class.