In class BilinearForm, add new methods, UseSparsity, that allocate the internal SparseMatrix using a given sparsity, either as CSR (I,J) arrays or as a SparseMatrix.
In class SparseMatrix, add new method, MoveDiagonalFirst, that moves the
diagonal entry in each row to the first position in the row and shifts the
entries that were before the diagonal up by one - this way the order of all
other entries is preserved.

Add some doxygen documentation for class UMFPackSolver.