Added a DG option to example 2, just the serial one. This required a few additions:
One addition is a new option for boundary face intergrators in BilinearForm: the
ability to integrate over a subset of the boundary. The option I added (to a new
version of BilinearForm::AddBdrFaceIntegrator) works with the same type of
essential boundary attribute arrays that we use in many of the examples.

Another addition is to make VectorBoundaryLFIntegrator work for DG spaces.