Fixed a small typo in a comment. Fixed the direction of the phi boundary condition which I think was accidentally changed.
Commented out a group of unused variables related to the L2 norm of the electric and magnetic fields and the electric losses.  I only commented these out because I like to see such things tracked (I think they were in an earlier version).  Did I miss a discussion about this?  Can we put this back in or was there an argument against this additional output?