Adds a function to DataCollectionUtility to truncate a field name when it ends with an integer.
The function strips an array index (when present) from a field name's string
(and decrements the index since the string name uses 1-based indexing).

This is the inverse of the GenerateFieldName function in the same class which optionally
appends an array_index (using 1-based indexing) to a given string.

Misc:
* Also adds a function to DataCollection to return the names of all registered grid functions