Rename class SundialsLinearSolver to SundialsODELinearSolver.
In the static wrapper methods Mult() and GradientMult() in class SundialsSolver,
the user_data has a pointer to the nonlinear Operator plus a pointer to the
computed Jacobian - this avoids unnecessary re-computation of the Jacobian, if
the point 'u' has not changed.

Remove the call to ARKodeSetOrder() in ARKODESolver::SetOrder().

In KinSolver::Mult() support the case iterative_mode = false.

Improve doxygen docs.