New interface for class TimeDependentOperator supporting explicitly and implicitly defined operators. See the class doxygen documentation for details.
In the INSTALL file, mention that MFEM can be built as a shared library with
CMake.

Add examples/petsc and examples/sundials to the doxygen documentation.

Add items for the sundials and petsc examples in CodeDocumentation.dox.

Rename class PetscSolverMonitorCtx to PetscSolverMonitor.

In class PetscParMatrix, rename SetData() to PlaceArray() and ResetData() to
ResetArray() - these names are taken from the corresponding PETSc functions that
are actually called.

Add several FIXME items in the source - these still need to be resolved.

Update the PETSc ex9p to use time stepping loop similar to the standard ex9p.

Add/tweak doxygen documentation in classes ODESolver, SparseMatrix, etc.

Other small tweaks in various places.